requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
pandas==2.0.3
swifter==1.3.4
selenium==4.10.0
numpy==1.23.5
